+++ NEWS.md
@@ -48,8 +48,10 @@
 
 * Remove occasional deprecation message coming from Polars when using `%in%`
   (#259, @ppanko).
-  
+
 * Better handling of functions prefixed with `<pkg>::` (#261).
+
+* Fix wrong behavior of `paste()` and `paste0()` with `collapse` (#263).
 
 ## Documentation